r_shadows fix
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@1348 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/gl/gl_ppl.c
@@ -119,8 +119,8 @@ typedef struct shadowmesh_s {
 struct {
 	short count;
 	short count2;
-	short next;
-	short prev;
+	int next;
+	int prev;
 } edge[MAX_MAP_EDGES];
 int firstedge;
